Move most of console.py to _pyodide package
